requirements.txt
Pillow
google-cloud-vision
pyperclip